// payload-types.ts
export interface Image {
  id: number;
  name?: string | null;
  _key?: string | null;
  prefix?: string | null;
  updatedAt: string;
  createdAt: string;
  url?: string | null;
  thumbnailURL?: string | null;
  filename?: string | null;
  mimeType?: string | null;
  filesize?: number | null;
  width?: number | null;
  height?: number | null;
  focalX?: number | null;
  focalY?: number | null;
}
/**
 * This interface was referenced by `Config`'s JSON-Schema
 * via the `definition` "posts".
 */
export interface Post {
  id: number;
  title: string;
  subtitle: string;
  featuredImage: number | Image;
  publishDate: string;
  author: {
    name: string;
    avatar: number | Image;
  };
  content: (
    | {
        title: string;
        id?: string | null;
        blockName?: string | null;
        blockType: 'titleBlock';
      }
    | {
        text: string;
        id?: string | null;
        blockName?: string | null;
        blockType: 'paragraphBlock';
      }
    | {
        heading?: string | null;
        text: string;
        id?: string | null;
        blockName?: string | null;
        blockType: 'textBlock';
      }
    | {
        image: number | Image;
        caption?: string | null;
        id?: string | null;
        blockName?: string | null;
        blockType: 'imageBlock';
      }
    | {
        quote: string;
        author?: string | null;
        id?: string | null;
        blockName?: string | null;
        blockType: 'quoteBlock';
      }
  )[];
  updatedAt: string;
  createdAt: string;
}